Finished authorization on remaining entities.

+++ LATESTReactKanBanDo/Controllers/ColumnController.cs
@@ -1,7 +1,12 @@
-﻿using LATESTReactKanBanDo.Data.Dtos;
+﻿using LATESTReactKanBanDo.Auth.Model;
+using LATESTReactKanBanDo.Data.Dtos;
 using LATESTReactKanBanDo.Data.Entities;
 using LATESTReactKanBanDo.Data.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Data;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
 
 namespace LATESTReactKanBanDo.Controllers
 {
@@ -43,8 +48,21 @@ namespace LATESTReactKanBanDo.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = KanbanRoles.KanbanUser)]
         public async Task<ActionResult<ColumnDto>> Create(int viewId, CreateColumnDto createColumnDto)
         {
+            var view = await _viewsRepository.GetAsync(viewId);
+            if (view == null)
+            {
+                return NotFound();
+            }
+
+            var userId = User.FindFirstValue(JwtRegisteredClaimNames.Sub);
+            if (view.UserId != userId)
+            {
+                return Forbid();
+            }
+
             var column = new Column { Name = createColumnDto.Name };
 
             await _columnsRepository.CreateAsync(column, viewId);
@@ -54,38 +72,61 @@ namespace LATESTReactKanBanDo.Controllers
         }
 
         [HttpPut]
-        [Route("{ColumnId}")]
-        public async Task<ActionResult<ColumnDto>> Update(int ColumnId, UpdateColumnDto updateColumnDto)
+        [Route("{columnId}")]
+        [Authorize(Roles = KanbanRoles.KanbanUser)]
+        public async Task<ActionResult<ColumnDto>> Update(int viewId, int columnId, UpdateColumnDto updateColumnDto)
         {
-            var column = await _columnsRepository.GetAsync(ColumnId);
+            var view = await _viewsRepository.GetAsync(viewId);
+            if (view == null)
+            {
+                return NotFound("View not found.");
+            }
 
+            var userId = User.FindFirstValue(JwtRegisteredClaimNames.Sub);
+            if (view.UserId != userId)
+            {
+                return Forbid();
+            }
+
+            var column = await _columnsRepository.GetAsync(columnId);
             if (column == null)
             {
-                return NotFound();
+                return NotFound("Column not found.");
             }
 
             column.Name = updateColumnDto.Name;
+
             await _columnsRepository.UpdateAsync(column);
 
             return Ok(new ColumnDto(column.Id, column.Name));
         }
 
-        // api/Columns/{ColumnId}
         [HttpDelete]
-        [Route("{ColumnId}")]
-        public async Task<ActionResult> Remove(int ColumnId)
+        [Route("{columnId}")]
+        [Authorize(Roles = KanbanRoles.KanbanUser)]
+        public async Task<ActionResult> Remove(int viewId, int columnId)
         {
-            var column = await _columnsRepository.GetAsync(ColumnId);
+            var view = await _viewsRepository.GetAsync(viewId);
+            if (view == null)
+            {
+                return NotFound("View not found.");
+            }
 
+            var userId = User.FindFirstValue(JwtRegisteredClaimNames.Sub);
+            if (view.UserId != userId)
+            {
+                return Forbid();
+            }
+
+            var column = await _columnsRepository.GetAsync(columnId);
             if (column == null)
             {
-                return NotFound();
+                return NotFound("Column not found.");
             }
 
             await _columnsRepository.DeleteAsync(column);
 
-            //204
-            return NoContent();
+            return NoContent(); // 204 No Content
         }
     }
 }
